+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 [LICENSE](LICENSE) - [README](README.md) - [CHANGELOG](CHANGELOG.md) - [TODO](TODO.md)
 
+## [0.1.2] - 12/05/2017 - [@emimontesdeoca](https://github.com/emimontesdeoca)
+### Add
+- Added atributes and methods to Elements.cs
+- Added atributes and methods to License.cs
+- Added atributes and methods to Template.cs
+- Added atributes and methods to Proyects.cs
+- Added atributes and methods to UserLicense.cs
+- Added atributes and methods to Elements.cs
+- Created Role
+- Added atributes and methods to Role.cs
+- Added atributes and methods to Users.cs
+
 ## [0.1.1] - 11/05/2017 - [@emimontesdeoca](https://github.com/emimontesdeoca)
 ### Added
 - Modified folder structure